src/idrac-sse/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
httpx
requests
httpx_sse
tenacity
stomp.py    
pandas